+++ services/value-stat-getter.js
@@ -29,6 +29,7 @@ function ValueStatGetter(model, params) {
         var field = filter.field;
         if (field.indexOf(':') !== -1) {
           field = '$' + field.replace(':', '.') + '$';
+          // TOFIX: the column name might be different than the model name.
         }
 
         var condition = {};