+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Resource Retrieval - Fix the record retrieval for a record having no belongsTo/hasOne for projects using Sequelize 4.8.x.
+- Charts - Fix broken charts if they contain relationship filters on model fields that have a different column name.
 
 ## RELEASE 1.3.6 - 2017-09-10
 ### Changed